Safari-609.1.11.1/LayoutTests/js/instanceof-operator-expected.txt

instanceof test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getterCalled is false
PASS foo() is false
PASS successfullyParsed is true

TEST COMPLETE
